requirements.txt
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
matplotlib>=3.7.0
neo4j>=5.14.0
sentence-transformers>=2.2.0
groq>=0.4.0
torch>=2.0.0
transformers>=4.36.0
tqdm>=4.65.0
python-dotenv>=1.0.0
accelerate>=0.27.0
bitsandbytes>=0.41.0
safetensors>=0.4.0 